overlays/osk: move pointer variables to fxo
